rename pages package to 'selenium.pageobjects', remove PageObject class, rename RHStartPageObject class to 'Start'
